Dial down on the abstraction in WC Model
